What if we just re-use code snippets in other docs?
