+++ homepage/homepage/content/docs/code-snippets/permissions-and-sharing/sharing/create-invite-link.svelte.ts
@@ -1,7 +1,10 @@
 //[!code hide]
 import { Organization } from "./schema";
 //[!code hide]
-const organization = Organization.create({ name: "Garden Computing" });
+const organization = Organization.create({
+  name: "Garden Computing",
+  projects: [],
+});
 import { createInviteLink } from "jazz-tools/svelte";
 
 const inviteLink = createInviteLink(organization, "writer");
