+++ homepage/homepage/content/docs/code-snippets/permissions-and-sharing/sharing/create-invite-link.ts
@@ -1,11 +1,16 @@
 //[!code hide]
-import { Organization } from "./schema.ts";
-//[!code hide]
-import { createJazzTestAccount } from "jazz-tools/testing";
-const me = await createJazzTestAccount({
-  isCurrentActiveAccount: true,
+import { JazzAccount, Organization } from "./schema.ts";
+const me = await JazzAccount.getMe().$jazz.ensureLoaded({
+  resolve: {
+    root: {
+      organizations: true,
+    },
+  },
 });
-const organization = Organization.create({ name: "Garden Computing" });
+const organization = Organization.create({
+  name: "Garden Computing",
+  projects: [],
+});
 
 // #region CreateLink
 import { createInviteLink } from "jazz-tools";
@@ -23,5 +28,9 @@ import { consumeInviteLink } from "jazz-tools";
 consumeInviteLink({
   inviteURL: inviteLink,
   invitedObjectSchema: Organization, // Pass the schema for the invited object
+}).then(async (invitedObject) => {
+  if (!invitedObject) throw new Error("Failed to consume invite link");
+  const organization = await Organization.load(invitedObject?.valueID);
+  me.root.organizations.$jazz.push(organization);
 });
 // #endregion
