+++ homepage/homepage/content/docs/code-snippets/permissions-and-sharing/sharing/expo.tsx
@@ -1,15 +1,29 @@
 import { co, z } from "jazz-tools";
 
-import { Organization } from "./schema";
-const organization = Organization.create({ name: "Garden Computing" });
+import { JazzAccount, Organization } from "./schema";
+const organization = Organization.create({
+  name: "Garden Computing",
+  projects: [],
+});
+
+const me = await JazzAccount.getMe().$jazz.ensureLoaded({
+  resolve: {
+    root: {
+      organizations: true,
+    },
+  },
+});
 
 // #region AcceptInvite
 import { useAcceptInviteNative } from "jazz-tools/expo";
 
 useAcceptInviteNative({
   invitedObjectSchema: Organization,
-  onAccept: (organizationID) => {
-    console.log("Accepted invite!");
+  onAccept: async (organizationID) => {
+    const organization = await Organization.load(organizationID);
+    if (!organization.$isLoaded)
+      throw new Error("Organization could not be loaded");
+    me.root.organizations.$jazz.push(organization);
     // navigate to the organization page
   },
 });
