+++ homepage/homepage/content/docs/code-snippets/permissions-and-sharing/sharing/react-snippet.tsx
@@ -1,5 +1,16 @@
-import { Organization } from "./schema";
-const organization = Organization.create({ name: "Garden Computing" });
+import { JazzAccount, Organization } from "./schema";
+const organization = Organization.create({
+  name: "Garden Computing",
+  projects: [],
+});
+
+const me = await JazzAccount.getMe().$jazz.ensureLoaded({
+  resolve: {
+    root: {
+      organizations: true,
+    },
+  },
+});
 
 // #region CreateInviteLink
 import { createInviteLink } from "jazz-tools/react";
@@ -12,8 +23,11 @@ import { useAcceptInvite } from "jazz-tools/react";
 
 useAcceptInvite({
   invitedObjectSchema: Organization,
-  onAccept: (organizationID) => {
-    console.log("Accepted invite!");
+  onAccept: async (organizationID) => {
+    const organization = await Organization.load(organizationID);
+    if (!organization.$isLoaded)
+      throw new Error("Organization could not be loaded");
+    me.root.organizations.$jazz.push(organization);
     // navigate to the organization page
   },
 });
